added combigrid example
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@5573 4eea3252-f0fb-4393-894d-40516dce545b

+++ combigrid/src/sgpp/combigrid/fullgrid/GridContainer.hpp
@@ -122,7 +122,7 @@ class FGridContainer {
                  const std::vector<bool>& hasBoundaryPts, _Tp coef,
                  BasisFunctionBasis* basis = NULL) {
 
-    int dim = levels.size();
+    int dim = static_cast<int>(levels.size());
 
     _fullgrid = new combigrid::FullGrid<_Tp>(dim, levels, hasBoundaryPts, basis);
     _fullgrid_levels = levels;
